+++ Engine/Source/Developer/ScreenShotComparisonTools/Public/ImageComparer.h
@@ -2,9 +2,18 @@
 
 #pragma once
 
+#include "Containers/ContainersFwd.h"
+#include "Containers/UnrealString.h"
 #include "CoreMinimal.h"
-#include "UObject/ObjectMacros.h"
+#include "Internationalization/Text.h"
+#include "Math/Color.h"
+#include "Math/UnrealMathSSE.h"
+#include "Misc/AssertionMacros.h"
+#include "Misc/DateTime.h"
 #include "Misc/Paths.h"
+#include "Templates/SharedPointer.h"
+#include "UObject/ObjectMacros.h"
+
 #include "ImageComparer.generated.h"
 
 class Error;
@@ -100,8 +109,8 @@ public:
 	{
 		const bool AlphaSimilar = FMath::IsNearlyEqual((float)ColorA.A, ColorB.A, Tolerance.Alpha);
 
-		const float BrightnessA = FPixelOperations::GetLuminance(ColorA);
-		const float BrightnessB = FPixelOperations::GetLuminance(ColorB);
+		const double BrightnessA = FPixelOperations::GetLuminance(ColorA);
+		const double BrightnessB = FPixelOperations::GetLuminance(ColorB);
 		const bool BrightnessSimilar = FMath::IsNearlyEqual(BrightnessA, BrightnessB, Tolerance.MinBrightness);
 
 		return BrightnessSimilar && AlphaSimilar;
@@ -126,8 +135,8 @@ public:
 
 	static FORCEINLINE bool IsContrasting(const FColor& ColorA, const FColor& ColorB, const FImageTolerance& Tolerance)
 	{
-		const float BrightnessA = FPixelOperations::GetLuminance(ColorA);
-		const float BrightnessB = FPixelOperations::GetLuminance(ColorB);
+		const double BrightnessA = FPixelOperations::GetLuminance(ColorA);
+		const double BrightnessB = FPixelOperations::GetLuminance(ColorB);
 
 		return FMath::Abs(BrightnessA - BrightnessB) > Tolerance.MaxBrightness;
 	}
