+++ Engine/Source/Editor/Blutility/Private/GlobalEditorUtilityBase.cpp
@@ -5,7 +5,7 @@
 #include "Misc/PackageName.h"
 #include "GameFramework/Actor.h"
 #include "Editor/EditorPerProjectUserSettings.h"
-#include "AssetData.h"
+#include "AssetRegistry/AssetData.h"
 #include "Engine/Selection.h"
 #include "Editor.h"
 #include "ScopedTransaction.h"
@@ -64,7 +64,7 @@ void UDEPRECATED_GlobalEditorUtilityBase::GetSelectionBounds(FVector& Origin, FV
 
 	Origin = Extents.Origin;
 	BoxExtent = Extents.BoxExtent;
-	SphereRadius = Extents.SphereRadius;
+	SphereRadius = (float)Extents.SphereRadius;
 }
 
 void UDEPRECATED_GlobalEditorUtilityBase::ForEachSelectedActor()
