+++ Engine/Source/Runtime/GeometryCore/Public/Util/RefCountVector.h
@@ -136,7 +136,7 @@ public:
 			int j = (int)RefCounts.GetLength();
 			while (j < Index)
 			{
-				int InvalidCount = INVALID_REF_COUNT;	// required on older clang because a constexpr can't be passed by ref
+				unsigned short InvalidCount = INVALID_REF_COUNT;	// required on older clang because a constexpr can't be passed by ref
 				RefCounts.Add(InvalidCount);
 				FreeIndices.Add(j);
 				++j;
@@ -178,7 +178,7 @@ public:
 			int j = (int)RefCounts.GetLength();
 			while (j < Index)
 			{
-				int InvalidCount = INVALID_REF_COUNT;	// required on older clang because a constexpr can't be passed by ref
+				unsigned short InvalidCount = INVALID_REF_COUNT;	// required on older clang because a constexpr can't be passed by ref
 				RefCounts.Add(InvalidCount);
 				++j;
 			}
@@ -237,7 +237,7 @@ public:
 	 * @param IncrementRefCount Callable for incrementing an item, i.e. an item that has been referenced before.
 	 */
 	template <typename IterateFunc, typename AllocateRefCountFunc, typename IncrementRefCountFunc>
-	void Rebuild(SIZE_T Num, IterateFunc&& Iterate, AllocateRefCountFunc&& AllocateRefCount, IncrementRefCountFunc&& IncrementRefCount)
+	void Rebuild(unsigned int Num, IterateFunc&& Iterate, AllocateRefCountFunc&& AllocateRefCount, IncrementRefCountFunc&& IncrementRefCount)
 	{
 		// Initialize ref counts to the given number of elements.
 		RefCounts.Resize(Num);
@@ -265,10 +265,10 @@ public:
 		Forward<IterateFunc>(Iterate)(UpdateRefCount);
 
 		// Add unused elements to free list.
-		const SIZE_T FreeIndicesNum = Num - UsedCount;
+		const unsigned int FreeIndicesNum = Num - UsedCount;
 		FreeIndices.SetNum(FreeIndicesNum);
-		SIZE_T FreeIndicesIndex = 0;
-		for (SIZE_T Index = 0; (Index < Num) & (FreeIndicesIndex < FreeIndicesNum); ++Index)
+		unsigned int FreeIndicesIndex = 0;
+		for (unsigned int Index = 0; (Index < Num) & (FreeIndicesIndex < FreeIndicesNum); ++Index)
 		{
 			if (RefCounts[Index] == INVALID_REF_COUNT)
 			{
@@ -595,8 +595,8 @@ public:
 			return false;
 		}
 
-		const int32 Num = FMath::Max(Lhs.GetMaxIndex(), Rhs.GetMaxIndex());
-		for (int32 Idx = 0; Idx < Num; ++Idx)
+		const size_t Num = FMath::Max(Lhs.GetMaxIndex(), Rhs.GetMaxIndex());
+		for (size_t Idx = 0; Idx < Num; ++Idx)
 		{
 			const bool LhsIsValid = Lhs.IsValid(Idx);
 			if (LhsIsValid != Rhs.IsValid(Idx))
